Move cloud function sources into src tree
