+++ src/cloud/process-new-page/index.js
