+++ src/cloud/process-new-story/index.js
