+++ src/cloud/render-variant/buildAltsHtml.js
