+++ src/cloud/submit-new-page/helpers.js
