+++ src/cloud/submit-new-page/index.js
