+++ test/cloud-functions/buildHtml.test.js
@@ -1,6 +1,6 @@
 import { describe, test, expect } from '@jest/globals';
 import { JSDOM } from 'jsdom';
-import { buildHtml } from '../../infra/cloud-functions/render-variant/buildHtml.js';
+import { buildHtml } from '../../src/cloud/render-variant/buildHtml.js';
 
 describe('buildHtml', () => {
   test('sets default head title when story title missing', () => {
